msm: watchdog: make watchdog a platform device
There are now several conditionals that control how the watchdog
behaves. We have described these in terms cpu_is and machine_is macros.
Instead, convert tunables to platform data so we can keep the driver
clean.

Additionally, since it is a platform device, we can now move the
suspending code to the suspend_noirq function, which leaves the watchdog
on for more of the suspend process.

Change-Id: Ieba2fa19c19ac6efa38d4031783878929a83f50b
Signed-off-by: Jeff Ohlstein <johlstei@codeaurora.org>
[jeff.carlyle@motorola.com: adapted from 7e66855636bcf7f86af54bd31e264af719924358]
Signed-off-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>