msm: qdss: save and restore debug registers across PC
Debug OS lock register is on the core (Krait) power domain and
so looses its contents on a power collapse. It's power on reset
state is locked and external debugger accesses are not allowed
while the debug OS lock is set.

This prevents JTag users from setting breakpoints after the first
Krait power collapse.

Saving and restoring debug registers across power collapse helps
preserve the breakpoints as well as clear the debug OS lock as
part of this procedure while coming out of power collapse.

CRs-Fixed: 310982
Change-Id: Icfccea17e556474241bf495885d9bb575d7571bf
Signed-off-by: Pratik Patel <pratikp@codeaurora.org>
[jeff.carlyle@motorola.com: adapted from fd6f56ae53891285a8a19c7682d40d7dc80a7f13]
Signed-off-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>