IKQCOMSD1-382: Remove DMA mappings when disabling endpoint
This change addresses a memory corruption issue seen when a USB cable was
unplugged and replugged to a host PC. The device controller driver was not
unmapping outstanding DMA requests on disconnect. On reconnect when the
function driver resubmitted its usb requests the device controller would
see the previous mapping present and neglect to create a new mapping. In
case the function had freed and reallocated their RX request buffer
subsequent transactions on the endpoint would cause data to be written
to the old address.

Change-Id: I048da47004539f25b2c2fc2ee5b9190df95047f3
Signed-off-by: Neil Patel <NPATEL@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/317634
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>