IKQCOMSD1-609: Add Support for 4.35 V Charging
This is Patch creates a architecture so that Charger and
BMS data can be sent from the board file.  MMI implementation
uses the 1 Wire mmi-battery data to locate the proper data.
This allows the Peak Voltage to be sent ot he charger driver
and enable 4.35 V Max Charge Voltage.  Also the concept of
step charging is included as it is needed for proper 4.35 V
charging.  Simple data for EB20 and a default BMS table are
now available as well.

Change-Id: I39bb717edf56859e4ec52130c2ce2c6180a5d945
Signed-off-by: w65789 <w65789@motorola.com>
Change-Id: I2da7952b5fa2ecea3d09cc66e1f1e96458bf7803
Reviewed-on: http://gerrit.pcs.mot.com/318633
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Igor Kovalenko <Igor.Kovalenko@motorola.com>

Conflicts:
	arch/arm/mach-msm/Makefile
	drivers/power/pm8921-bms.c
	drivers/power/pm8921-charger.c
	include/linux/mfd/pm8xxx/pm8921-bms.h
	include/linux/mfd/pm8xxx/pm8921-charger.h