IKVANQUISH-255:  Touchscreen does not work after display timeout
Since the Touch IRQ (GPIO 46) is an open drain signal being driven
by the Melfas touch IC, and an external pull up is placed on the PCB,
SW MUST disable any internal pull down's inside the MSM.  Not doing
so will cause approximately 1/2 of the IRQ voltage to be present on
the MSM pin.  This will lead to interrupt storms, poor system
performance, and a current drain hit.

Disabling the internal pull-down on GPIO 46 must occur in 2 locations:

1 - ./vendor/qcom/nonhlos/boot_images/core/dal/config/boot_ssbl/TLMM.xml
2 - ./kernel/arch/arm/mach-msm/board-mmi-melfas-touch.c

TLMM.xml is used to configure GPIO's / pins during sleep.
board-mmi-melfas-touch.c is used by the kernel driver for pin init.

Change-Id: If4ba9eb05c799f615300618b4ab0265486806b56
Reviewed-on: http://gerrit.pcs.mot.com/323612
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Igor Kovalenko <Igor.Kovalenko@motorola.com>