IKQCOMSD1-539: Add Support for Battery Temperature Charging Thresholds
This Change Implements both the Extended Range and the
Out of Temperature Ranges.  The Battery Can only be
Charged to 3.8 V in the Extended Range and can not be
Charged at all in the Out of Temperature Range.  Also
If the Phone escapes 10 degrees C in either hte Hot or
Cold direction past the Out Thresholds the Phone will
Shutdown.

Change-Id: Ie526b5b4673fe3272e17aea22c1e1fb911fa9c8a
Signed-off-by: w65789 <w65789@motorola.com>
Change-Id: I5d3dced484636720d7f5e41b12ac47ce891ed62e
Reviewed-on: http://gerrit.pcs.mot.com/325010
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>

Conflicts:
	drivers/power/pm8921-charger.c