IKQCOMSD1-880: msm: enable softlockup lockup panic
Enable softlockup panics.  The kernel hard lockup timer is 10s, which
sets the soft lockup timer to 20s.  The MSM watchdog timer has to be
set higher than this, since it uses a work queue to pet the dog.
Presumably if the soft lockup detection is going to go off, the work
queue will also not be serviced.

Change-Id: I0a07e0b345acc996763837191562b46f7cc58d56
Signed-off-by: Russ Knize <Russ.Knize@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/339982
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Russell Knize <russ.knize@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Mark Melone <g15829@motorola.com>