msm: dsps: Revise handling of PPSS_PAUSE register.
Only modify the PPSS_PAUSE register if the apps processor controls
the DSPS power state. The Linux kernel shouldn't modify the register
if the DSPS is in control of its own power state.

Also move the definition of the PPSS_PAUSE_REG address to the devices
file, since it may be different on different platforms.

Change-Id: Ief71e57d850d45ecf5e03863e3c37d8d485cd6ff
CR-fixed: 370236
Signed-off-by: Karthik Karuppasamy <kkaruppa@codeaurora.org>
Signed-off-by: Jason Hrycay <jason.hrycay@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/442963
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Jagatkumar Shah <w18764@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Christopher Fries <C.Fries@motorola.com>
Reviewed-by: Igor Kovalenko <Igor.Kovalenko@motorola.com>