IKHSS7-41365: Synchronize VFE starts/stops to prevent kernel panics
VFE starts and stops are not atomic and they are not synchronous operations so
one thread stopping VFE enters this section of code and polls the AXI status
register until AXI is not busy while another thread enters VFE to start up
which involves initializing AXI. If one thread is waiting on AXI status while
another resets/inits AXI the status reads will fail - especially if the AXI bus
mapping should change in a subsequent init cycle.

This is a QC patch to solve the problem mentioned above.

Change-Id: I79f32c7c3c24fb4d8a828ab4822d8a9569babf88
Signed-off-by: Rob Kubicki <w17311@motorola.com>